First attempt at testing the async API
We want to make sure that responses that don't take the same time to execute still are returned to the proper request
